requirements.txt
numpy>=1.20.0
jax==0.2.13
psi4==1.3.2+ecbda83
tequila-basic==1.5.1